fix: scp_get, scp_putのファイルサイズの型を64ビット整数に

+++ tobas_gui/tobas_gui_common/include/tobas_gui_common/ssh_client.hpp
@@ -32,7 +32,7 @@ public:
   Impl::Error scpGet(
     const std::string& remote_path,
     const std::string& local_path,
-    std::function<void(uint32_t, uint32_t)> callback = nullptr);
+    std::function<void(uint64_t, uint64_t)> callback = nullptr);
 
   Impl::Error scpPut(
     const std::string& local_dir,
@@ -40,7 +40,7 @@ public:
     bool parents,
     const std::vector<std::string>& exclude_dirs,
     bool superuser = false,
-    std::function<void(uint32_t, uint32_t)> callback = nullptr);
+    std::function<void(uint64_t, uint64_t)> callback = nullptr);
 
   Impl::Error sftpRead(const std::string& remote_path, std::string& text, bool superuser = false);
